requirements.txt
charset-normalizer==3.1.0
cycler==0.11.0
fonttools==4.38.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.6.0
kiwisolver==1.4.4
largestinteriorrectangle==0.2.0
llvmlite==0.39.1
matplotlib==3.5.3
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.31.0
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
stitching==0.4.0
tifffile==2021.11.2
timm==0.5.4
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.3
zipp==3.15.0